Moving context objects into config folder
