+++ src/python/PythonSDK/PythonSDK.pyproj
@@ -22,9 +22,8 @@
     <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="foundationallm\context\context.py" />
-    <Compile Include="foundationallm\context\user_identity.py" />
-    <Compile Include="foundationallm\context\__init__.py" />
+    <Compile Include="foundationallm\config\context.py" />
+    <Compile Include="foundationallm\config\user_identity.py" />
     <Compile Include="foundationallm\models\list_option.py" />
     <Compile Include="foundationallm\resolvers\generic_resolver_agent.py" />
     <Compile Include="foundationallm\hubs\agent\agent_type.py" />
@@ -174,7 +173,6 @@
     <Folder Include="foundationallm\models\metadata\" />
     <Folder Include="foundationallm\models\orchestration\" />
     <Folder Include="foundationallm\hubs\" />
-    <Folder Include="foundationallm\context\" />
     <Folder Include="foundationallm\resolvers\" />
     <Folder Include="foundationallm\storage\" />
   </ItemGroup>
