+++ src/python/PythonSDK/foundationallm/config/__init__.py
@@ -1,5 +1,6 @@
 """ 
-Configuration methods for FoundationAL LM Python SDK
+Configuration classes for FoundationaLLM Python SDK
 """
-
 from .configuration import Configuration
+from .user_identity import UserIdentity
+from .context import Context
