--- src/python/PythonSDK/foundationallm/context/__init__.py
@@ -1,2 +1,0 @@
-from .user_identity import UserIdentity
-from .context import Context
