+++ src/python/PythonSDK/foundationallm/hubs/prompt/prompt_resolver.py
@@ -1,6 +1,6 @@
+from foundationallm.config import Context
 from foundationallm.hubs.prompt import PromptHubRequest, PromptHubResponse
 from foundationallm.hubs import Resolver
-from foundationallm.context import Context
 
 class PromptResolver(Resolver):
     """The PromptResolver class is responsible for resolving a request to a metadata value."""
